test: move common function
