+++ tests/integration_classes.py
@@ -7,6 +7,13 @@ except ImportError:
     import mock
 
 
+def _split_cmd_args(cmdargs):
+    if not cmdargs:
+        return []
+    else:
+        return cmdargs.split(' ')
+
+
 class MyApplication(smclip.CommandGroup):
 
     def __init__(self):
@@ -75,6 +82,7 @@ class PreprocessingCommand(smclip.Command):
 
     def add_arguments(self, parser):
         parser.add_argument('--toreplace')
+
 
 class ItemGroupCommand(smclip.CommandGroup):
     """Print help"""
